catalog

+++ Backend/Services/Bang/Bang.BLL/Application/Commands/Character/Commands/UpdateCharacterCommand.cs
@@ -6,10 +6,10 @@ namespace Bang.BLL.Application.Commands.Commands
 {
     public class UpdateCharacterCommand : IRequest
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public CharacterDto Dto { get; set; }
 
-        public UpdateCharacterCommand(long id, CharacterDto dto)
+        public UpdateCharacterCommand(int id, CharacterDto dto)
         {
             Id = id;
             Dto = dto;
